requirements.txt
streamlit
google-generativeai
arxiv
requests